+++ main/init/package.mill
@@ -8,7 +8,7 @@ import mill.util.Jvm
 
 object `package` extends RootModule with build.MillPublishScalaModule {
 
-  def moduleDeps = Seq(build.main, build.main.init.buildgen, build.scalalib)
+  def moduleDeps = Seq(build.main, build.scalalib)
 
   override def resources = Task {
     super.resources() ++ Seq(exampleList())
@@ -66,7 +66,7 @@ object `package` extends RootModule with build.MillPublishScalaModule {
   }
 
   object buildgen extends build.MillPublishScalaModule {
-    def moduleDeps = Seq(build.core.internal, build.main.util)
+    def moduleDeps = Seq(build.main.init, build.core.internal, build.main.util)
     def testModuleDeps = super.testModuleDeps ++ Seq(build.scalalib)
 
     // I tried moving `Tree` into this module, but it doesn't compile with Scala 2.12.20.
