+++ main/init/src/mill/init/BuildGenModule.scala
@@ -2,7 +2,6 @@ package mill.init
 
 import mill.define.{PathRef}
 import mill.api.{Result}
-import mill.main.buildgen.BuildGenUtil
 import mill.scalalib.{CoursierModule, Dep}
 import mill.scalalib.scalafmt.ScalafmtWorkerModule
 import mill.util.Jvm
@@ -21,7 +20,7 @@ trait BuildGenModule extends CoursierModule with TaskModule {
 
   def buildGenMainClass: T[String]
 
-  def buildGenScalafmtConfig: T[PathRef] = PathRef(BuildGenUtil.scalafmtConfigFile)
+  def buildGenScalafmtConfig: T[PathRef] = PathRef(mill.init.Util.scalafmtConfigFile)
 
   def init(args: String*): Command[Unit] = Task.Command(exclusive = true) {
     val root = moduleDir
@@ -38,7 +37,7 @@ trait BuildGenModule extends CoursierModule with TaskModule {
     ).exitCode
 
     if (exitCode == 0) {
-      val files = BuildGenUtil.buildFiles(root).map(PathRef(_)).toSeq
+      val files = mill.init.Util.buildFiles(root).map(PathRef(_)).toSeq
       val config = buildGenScalafmtConfig()
       Task.log.info("formatting Mill build files")
       ScalafmtWorkerModule.worker().reformat(files, config)
